requirements.txt
sentence-transformers 
langchainhub
pinecone-client[grpc] 
langchain 
langchain_community 
unstructured 
langchain-text-splitters 
tiktoken
pypdf
groq
sentence-transformers
langchain-pinecone
fastapi
ctransformers
python-multipart
aiofiles
